pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyplayer"
version = "0.6.0"
description = "Powerful video player and editor built on VLC and PyQt5"
requires-python = ">=3.13"
dependencies = [
    "PyQt5>=5.15.9",
    "python-vlc>=3.0.18122",
    "pillow>=9.5.0",
    "requests>=2.31.0",
    "pywin32>=306;sys_platform=='win32'",
    "colour>=0.1.5",
    "filetype>=1.0.13",
    "tinytag>=1.9.0",
    "music-tag>=0.4.3",
    "Send2Trash>=1.8.2",
]

[project.gui-scripts]
pyplayer = "pyplayer.app:main"

[project.optional-dependencies]
build = ["pyinstaller>=5.13.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pyplayer = ["py.typed"]
